refactor: Remove index in features/main, Change import statement
