+++ src/features/index.js
@@ -1,4 +1,3 @@
 import { ShoutGoContainer, ShoutGo } from "./shoutGo";
-import { MainContainer, Main, Page } from "./main";
 
-export { MainContainer, Main, ShoutGoContainer, ShoutGo, Page };
+export { ShoutGoContainer, ShoutGo };
