--- src/features/main/index.js
@@ -1,4 +1,0 @@
-import MainContainer from "./container/MainContainer";
-import { GameGuide, Main, Page } from "./presenter";
-
-export { MainContainer, GameGuide, Main, Page };
