--- src/features/main/presenter/index.js
@@ -1,5 +1,0 @@
-import GameGuide from "./GameGuide";
-import Main from "./Main";
-import Page from "./Page";
-
-export { GameGuide, Main, Page };
